feat: Add modals for task deletion and image upload, enhance task management with drag-and-drop functionality feat: Add modals for task deletion and image upload, enhance task management with drag-and-drop functionality
- Implemented DeleteTaskModal for confirming task deletions.
- Created ImageUpload component for handling image uploads with preview and removal options.
- Added PrioritySelect component for selecting task priority with visual indicators.
- Developed TaskCard component to display individual tasks with edit and delete options.
- Introduced TaskColumn component to manage task organization and addition.
- Enhanced board management with useBoardManagement and useBoardState hooks for state management.
- Implemented useDragAndDrop hook for drag-and-drop functionality between task columns.
- Created useProjectFilters and useProjectQuery hooks for managing project-related data.
- Developed useTaskManagement and useTaskMutations hooks for handling task operations.
- Integrated socket functionality for real-time task updates with useTaskSocket.
- Added DueDateStatus component to display task due date status visually.
