+++ packages/Frontend/src/component/pages/Projects/AddProjectTask.jsx
@@ -1,23 +1,7 @@
 import { useEffect, useState } from "react";
-import {
-  Ban,
-  Calendar,
-  CircleAlert,
-  SignalHigh,
-  SignalLow,
-  SignalMedium,
-  Trash,
-  Upload,
-  X,
-} from "lucide-react";
-import {
-  Button,
-  CircularProgress,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-} from "@mui/material";
+import { Button } from "@mui/material";
+import ImageUpload from "./ProjectModals/ImageUpload";
+import PrioritySelect from "./ProjectModals/PrioritySelect";
 
 const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
   const [status, setStatus] = useState("Pending");
@@ -33,7 +17,6 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
   const [imagePreview, setImagePreview] = useState(
     editTask?.image?.[0]?.url || editTask?.image || null
   );
-  const [isDeletingImage, setIsDeletingImage] = useState(false);
 
   useEffect(() => {
     if (editTask) {
@@ -56,28 +39,14 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
     }
   }, [editTask]);
 
-  const handleFileChange = (e) => {
-    const file = e.target.files[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        setImagePreview(reader.result);
-        setImageFile(file);
-      };
-      reader.readAsDataURL(file);
-    }
+  const handleImageChange = (preview, file) => {
+    setImagePreview(preview);
+    setImageFile(file);
   };
 
-  const handleRemoveImage = () => {
-    setIsDeletingImage(true);
-    try {
-      setImagePreview(null);
-      setImageFile(null);
-    } catch (error) {
-      console.error("Delete image error:", error);
-    } finally {
-      setIsDeletingImage(false);
-    }
+  const handleImageRemove = () => {
+    setImagePreview(null);
+    setImageFile(null);
   };
 
   const createFormData = () => {
@@ -91,7 +60,6 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
     if (imageFile instanceof File) {
       formData.append("image", imageFile);
     } else if (editTask && imagePreview && !imageFile) {
-      // If editing and there's an existing image, send the existing image data
       const existingImage = editTask.image?.[0];
       if (existingImage) {
         formData.append("image", JSON.stringify([existingImage]));
@@ -130,49 +98,13 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
           </h2>
         </div>
         <form onSubmit={handleSubmit}>
-          <div className="flex flex-col justify-center items-center mb-4">
-            {imagePreview ? (
-              <div className="relative mb-3">
-                <img
-                  src={imagePreview}
-                  alt="Task preview"
-                  className="rounded-3xl h-40 w-full object-cover"
-                />
-                <button
-                  type="button"
-                  onClick={handleRemoveImage}
-                  className="absolute -right-8 bottom-1 p-1 rounded-full"
-                  disabled={loading || isDeletingImage}
-                >
-                  {isDeletingImage ? (
-                    <CircularProgress size={24} sx={{ color: "#dc2626" }} />
-                  ) : (
-                    <Trash className="w-5 h-5 text-red-600" />
-                  )}
-                </button>
-              </div>
-            ) : null}
+          <ImageUpload
+            imagePreview={imagePreview}
+            onImageChange={handleImageChange}
+            onImageRemove={handleImageRemove}
+            loading={loading}
+          />
 
-            <label
-              htmlFor="taskImage"
-              className={`flex gap-2 p-2 justify-center cursor-pointer border border-dashed border-gray-400 px-4 py-3 rounded-[10px] bg-[#f8f8f8] mb-1 text-border font-medium ${
-                loading || isDeletingImage
-                  ? "text-gray-400 cursor-not-allowed"
-                  : "text-gray-700"
-              }`}
-            >
-              Upload image
-              <Upload />
-            </label>
-            <input
-              type="file"
-              className="hidden"
-              id="taskImage"
-              name="image"
-              onChange={handleFileChange}
-              disabled={loading || isDeletingImage}
-            />
-          </div>
           <div className="mb-3">
             <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Task Name
@@ -186,6 +118,7 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
               required
             />
           </div>
+
           <div className="mb-3">
             <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Description
@@ -198,6 +131,7 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
               rows="4"
             />
           </div>
+
           <div className="mb-3">
             <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Status
@@ -221,119 +155,14 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
               </option>
             </select>
           </div>
+
           <div className="mb-6">
-            <FormControl fullWidth>
-              <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium mb-1">
-                Priority
-              </label>
-              <InputLabel id="status-label">Priority</InputLabel>
-              jsx
-              <Select
-                labelId="status-label"
-                id="status-select"
-                value={priority}
-                label="Priority"
-                className="dark:!bg-[#2D2D2D] dark:border-0 dark:border-gray-500 dark:text-gray-300"
-                onChange={(e) => setPriority(e.target.value)}
-                MenuProps={{
-                  PaperProps: {
-                    className: "dark:bg-[#2D2D2D] dark:text-gray-300",
-                  },
-                }}
-                renderValue={(value) => {
-                  const iconMap = {
-                    Urgent: (
-                      <CircleAlert
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-red-600/20 text-red-500 border-red-600"
-                      />
-                    ),
-                    High: (
-                      <SignalHigh
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-orange-500/20 text-orange-500 border-orange-500"
-                      />
-                    ),
-                    Medium: (
-                      <SignalMedium
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-yellow-500/20 text-yellow-500 border-yellow-500"
-                      />
-                    ),
-                    Low: (
-                      <SignalLow
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-blue-500 text-blue-300 border-blue-300"
-                      />
-                    ),
-                    Normal: (
-                      <Ban
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-gray-500 text-gray-200 border-gray-300"
-                      />
-                    ),
-                  };
-                  return (
-                    <span className="flex dark:text-gray-300 items-center gap-2">
-                      {iconMap[value]} {value}
-                    </span>
-                  );
-                }}
-              >
-                <MenuItem
-                  className="dark:text-gray-300 dark:bg-[#2D2D2D] hover:dark:bg-[#404040]"
-                  value="Urgent"
-                >
-                  <CircleAlert
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-red-600/20 text-red-500 border-red-600"
-                  />
-                  Urgent
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="High"
-                >
-                  <SignalHigh
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-orange-500/20 text-orange-500 border-orange-500"
-                  />{" "}
-                  High
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Medium"
-                >
-                  <SignalMedium
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-yellow-500/20 text-yellow-500 border-yellow-500"
-                  />{" "}
-                  Medium
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Low"
-                >
-                  <SignalLow
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-blue-500 text-blue-300 border-blue-300"
-                  />{" "}
-                  Low
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Normal"
-                >
-                  <Ban
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-gray-500 text-gray-200 border-gray-300"
-                  />{" "}
-                  Normal
-                </MenuItem>
-              </Select>
-              {/* Optional: you can add helper text here if needed */}
-            </FormControl>
+            <PrioritySelect
+              priority={priority}
+              onPriorityChange={setPriority}
+            />
           </div>
+
           <div className="mb-3">
             <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Due Date
@@ -346,6 +175,7 @@ const AddProjectTask = ({ closeModal, onAddTask, editTask }) => {
               required
             />
           </div>
+
           <div className="flex justify-end mt-4">
             <Button
               type="submit"
