+++ packages/Frontend/src/component/pages/User_Tasks/AddTask.jsx
@@ -1,78 +1,57 @@
-import React, { useState } from "react";
-import {
-  Ban,
-  Calendar,
-  CircleAlert,
-  SignalHigh,
-  SignalLow,
-  SignalMedium,
-  Trash,
-  Upload,
-  X,
-} from "lucide-react";
-import {
-  Button,
-  CircularProgress,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-} from "@mui/material";
+import { useEffect, useState } from "react";
+import { Button } from "@mui/material";
+import ImageUpload from "../Projects/ProjectModals/ImageUpload";
+import PrioritySelect from "../Projects/ProjectModals/PrioritySelect";
 
-const AddTask = ({
-  closeModal,
-  onAddTask,
-  initialStatus,
-  initialData,
-  isEditing,
-  disableStatus,
-}) => {
-  const [title, setTitle] = useState(initialData?.title || "");
-  const [description, setDescription] = useState(
-    initialData?.description || ""
-  );
-  const [dueDate, setDueDate] = useState(
-    initialData?.dueDate?.split("T")[0] || ""
-  );
-  const [status, setStatus] = useState(initialStatus || "Pending");
-  const [priority, setPriority] = useState("Normal");
+const AddTask = ({ closeModal, onAddTask, editTask }) => {
+  const [status, setStatus] = useState("Pending");
+  const [taskName, setTaskName] = useState("");
+  const [dueDate, setDueDate] = useState("");
+  const [description, setDescription] = useState("");
   const [loading, setLoading] = useState(false);
-
+  const [priority, setPriority] = useState("Normal");
+  const [disablePriority, setDisablePriority] = useState(false);
+  
   // States for image handling
   const [imageFile, setImageFile] = useState(null);
   const [imagePreview, setImagePreview] = useState(
-    initialData?.image?.[0]?.url || initialData?.image || null
+    editTask?.image?.[0]?.url || editTask?.image || null
   );
-  const [isDeletingImage, setIsDeletingImage] = useState(false);
 
-  const handleFileChange = (e) => {
-    const file = e.target.files[0];
-    if (file) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        setImagePreview(reader.result);
-        setImageFile(file);
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-
-  const handleRemoveImage = () => {
-    setIsDeletingImage(true);
-    try {
+  useEffect(() => {
+    if (editTask) {
+      setTaskName(editTask.title || "");
+      setDescription(editTask.description || "");
+      setDueDate(editTask.dueDate ? editTask.dueDate.substring(0, 10) : "");
+      setStatus(editTask.status || "Pending");
+      setPriority(editTask.priority || "Normal");
+      setDisablePriority(!!editTask.priority);
+      setImagePreview(editTask.image?.[0]?.url || editTask.image || null);
+    } else {
+      setTaskName("");
+      setDescription("");
+      setDueDate("");
+      setStatus("Pending");
+      setPriority("Normal");
+      setDisablePriority(false);
       setImagePreview(null);
       setImageFile(null);
-      toast.success("Image removed successfully");
-    } catch (error) {
-      toast.error("Failed to remove image");
-      console.error("Delete image error:", error);
-    } finally {
-      setIsDeletingImage(false);
     }
+  }, [editTask]);
+
+  const handleImageChange = (preview, file) => {
+    setImagePreview(preview);
+    setImageFile(file);
   };
+
+  const handleImageRemove = () => {
+    setImagePreview(null);
+    setImageFile(null);
+  };
+
   const createFormData = () => {
     const formData = new FormData();
-    formData.append("title", title || "");
+    formData.append("title", taskName || "");
     formData.append("description", description || "");
     if (dueDate) formData.append("dueDate", new Date(dueDate).toISOString());
     formData.append("status", status || "Pending");
@@ -80,15 +59,21 @@ const AddTask = ({
 
     if (imageFile instanceof File) {
       formData.append("image", imageFile);
-    } else if (isEditing && imagePreview && !imageFile) {
-      formData.append("image", "keep");
+    } else if (editTask && imagePreview && !imageFile) {
+      const existingImage = editTask.image?.[0];
+      if (existingImage) {
+        formData.append("image", JSON.stringify([existingImage]));
+      }
     }
 
     return formData;
   };
+
   const handleSubmit = async (e) => {
     e.preventDefault();
     setLoading(true);
+
+    if (!taskName.trim()) return;
 
     try {
       const formData = createFormData();
@@ -99,118 +84,62 @@ const AddTask = ({
   };
 
   return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 z-50 flex items-center justify-center">
-      <div className="bg-white dark:bg-[#080808]  shadow-lg p-6 w-[700px] relative !rounded-xl">
-        {/* <div className="bg-white w-[426px] rounded-[10px] shadow-lg p-6"> */}
+    <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
+      <div className="bg-white dark:bg-[#080808] shadow-lg p-6 w-[700px] relative !rounded-xl">
         <div className="flex justify-between items-center mb-4">
-          <h2 className="text-gray-500 dark:text-gray-400 text-2xl font-medium">
-            {isEditing ? "Edit Task" : "Add New Task"}
-          </h2>
           <button
             onClick={closeModal}
-            className="text-gray-400 hover:text-gray-600"
+            className="absolute top-2 right-2 text-gray-500 hover:text-gray-700 mt-4 mr-4"
           >
-            <X size={24} />
+            ✖
           </button>
+          <h2 className="text-xl dark:text-gray-400 font-semibold mb-4">
+            {editTask ? "Edit Task" : "Add New Task"}
+          </h2>
         </div>
-
         <form onSubmit={handleSubmit}>
-          <div className="flex flex-col justify-center items-center mb-4">
-            {imagePreview ? (
-              <div className="relative mb-3">
-                <img
-                  src={imagePreview}
-                  alt="Task preview"
-                  className="rounded-3xl h-40 w-full object-cover"
-                />
-                <button
-                  type="button"
-                  onClick={handleRemoveImage}
-                  className="absolute -right-8 bottom-1 p-1 rounded-full"
-                  disabled={loading || isDeletingImage}
-                >
-                  {isDeletingImage ? (
-                    <CircularProgress size={24} sx={{ color: "#dc2626" }} />
-                  ) : (
-                    <Trash className="w-5 h-5 text-red-600" />
-                  )}
-                </button>
-              </div>
-            ) : null}
+          <ImageUpload
+            imagePreview={imagePreview}
+            onImageChange={handleImageChange}
+            onImageRemove={handleImageRemove}
+            loading={loading}
+          />
 
-            <label
-              htmlFor="taskImage"
-              className={`flex gap-2 p-2 justify-center cursor-pointer border border-dashed border-gray-400 px-4 py-3 rounded-[10px] bg-[#f8f8f8] mb-1 text-border font-medium ${
-                loading || isDeletingImage
-                  ? "text-gray-400 cursor-not-allowed"
-                  : "text-gray-700"
-              }`}
-            >
-              Upload image
-              <Upload />
-            </label>
-            <input
-              type="file"
-              className="hidden"
-              id="taskImage"
-              name="image"
-              onChange={handleFileChange}
-              disabled={loading || isDeletingImage}
-            />
-          </div>
-
-          <div className="mb-4">
-            <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium mb-1">
-              Title
+          <div className="mb-3">
+            <label className="block dark:text-gray-400 text-sm font-medium mb-1">
+              Task Name
             </label>
             <input
               type="text"
-              value={title}
-              onChange={(e) => setTitle(e.target.value)}
-              className="w-full px-3 dark:bg-[#2D2D2D] dark:border-gray-500 dark:text-gray-300 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+              value={taskName}
+              onChange={(e) => setTaskName(e.target.value)}
+              placeholder="Task Name"
+              className="w-full p-2 dark:bg-[#2D2D2D] dark:text-gray-300 border border-gray-300 rounded"
               required
-              disabled={loading || isDeletingImage}
             />
           </div>
 
-          <div className="mb-4">
-            <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium mb-1">
+          <div className="mb-3">
+            <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Description
             </label>
             <textarea
               value={description}
               onChange={(e) => setDescription(e.target.value)}
-              className="w-full px-3 dark:bg-[#2D2D2D] dark:border-gray-500 dark:text-gray-300 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-              rows="3"
+              placeholder="Task Description"
+              className="w-full p-2 dark:bg-[#2D2D2D] dark:text-gray-300 border border-gray-300 rounded"
+              rows="4"
             />
           </div>
 
-          <div className="mb-4">
-            <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium mb-1">
-              Due Date
-            </label>
-            <div className="relative">
-              <input
-                type="date"
-                value={dueDate}
-                required
-                onChange={(e) => setDueDate(e.target.value)}
-                className="w-full dark:bg-[#2D2D2D] dark:border-gray-500 dark:text-gray-300 px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-              />
-            </div>
-          </div>
-
-          <div className="mb-6">
-            <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium mb-1">
+          <div className="mb-3">
+            <label className="block dark:text-gray-400 text-sm font-medium mb-1">
               Status
             </label>
             <select
               value={status}
               onChange={(e) => setStatus(e.target.value)}
-              className={`w-full px-3 py-2 dark:bg-[#2D2D2D] dark:border-gray-500 dark:text-gray-300 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 ${
-                disableStatus ? "bg-gray-100 text-gray-500" : ""
-              }`}
-              disabled={disableStatus}
+              className="w-full p-2 dark:bg-[#2D2D2D] border dark:text-gray-300 border-gray-300 rounded"
             >
               <option className="dark:text-gray-400" value="Pending">
                 Pending
@@ -222,130 +151,33 @@ const AddTask = ({
                 In Progress
               </option>
               <option className="dark:text-gray-400" value="Completed">
-                Completed
+                Done
               </option>
             </select>
           </div>
+
           <div className="mb-6">
-            <FormControl fullWidth>
-              <label className="block dark:text-gray-400 text-gray-700 text-sm font-medium !mb-2">
-                Priority
-              </label>
-              <InputLabel id="status-label ">Priority</InputLabel>{" "}
-              {/* Changed label to Priority */}
-              <Select
-                labelId="status-label"
-                id="status-select"
-                value={priority}
-                label="Priority"
-                className="dark:!bg-[#2D2D2D] dark:border-0 dark:border-gray-500 dark:text-gray-300"
-                onChange={(e) => setPriority(e.target.value)}
-                MenuProps={{
-                  PaperProps: {
-                    className: "dark:bg-[#2D2D2D] dark:text-gray-300",
-                  },
-                }}
-                renderValue={(value) => {
-                  const iconMap = {
-                    Urgent: (
-                      <CircleAlert
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-red-600/20 text-red-500 border-red-600"
-                      />
-                    ),
-                    High: (
-                      <SignalHigh
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-orange-500/20 text-orange-500 border-orange-500"
-                      />
-                    ),
-                    Medium: (
-                      <SignalMedium
-                        size={18}
-                        className=" mr-1 flex items-center  justify-center border rounded p-0.5 flex-shrink-0 bg-yellow-500/20 text-yellow-500 border-yellow-500"
-                      />
-                    ),
-                    Low: (
-                      <SignalLow
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-blue-500 text-blue-300 border-blue-300"
-                      />
-                    ),
-                    Normal: (
-                      <Ban
-                        size={18}
-                        className=" mr-1 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-gray-500 text-gray-200 border-gray-300"
-                      />
-                    ),
-                  };
-                  return (
-                    <span className="flex dark:text-gray-300 items-center gap-2">
-                      {iconMap[value]} {value}
-                    </span>
-                  );
-                }}
-              >
-                <MenuItem
-                  className="dark:text-gray-300 dark:bg-[#2D2D2D] hover:dark:bg-[#404040]"
-                  value="Urgent"
-                >
-                  <CircleAlert
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-red-600/20 text-red-500 border-red-600"
-                  />
-                  Urgent
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="High"
-                >
-                  <SignalHigh
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-orange-500/20 text-orange-500 border-orange-500"
-                  />{" "}
-                  High
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Medium"
-                >
-                  <SignalMedium
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-yellow-500/20 text-yellow-500 border-yellow-500"
-                  />{" "}
-                  Medium
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Low"
-                >
-                  <SignalLow
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-blue-500 text-blue-300 border-blue-300"
-                  />{" "}
-                  Low
-                </MenuItem>
-                <MenuItem
-                  className="dark:text-gray-400 dark:bg-[#2D2D2D] dark:border-gray-500"
-                  value="Normal"
-                >
-                  <Ban
-                    size={18}
-                    className="mr-2 flex items-center justify-center border rounded p-0.5 flex-shrink-0 bg-gray-500 text-gray-200 border-gray-300"
-                  />{" "}
-                  Normal
-                </MenuItem>
-              </Select>
-              {/* Optional: you can add helper text here if needed */}
-            </FormControl>
+            <PrioritySelect
+              priority={priority}
+              onPriorityChange={setPriority}
+              disabled={disablePriority}
+            />
           </div>
-          <div className="flex justify-end gap-4">
-            <Button
-              onClick={closeModal}
-              className="!text-base !capitalize !bg-[#7787e2] hover:shadow-lg hover:shadow-[#546FFF] !font-bold !text-white !py-3 !px-7 !rounded-xl"
-            >
-              Cancel
-            </Button>
+
+          <div className="mb-3">
+            <label className="block dark:text-gray-400 text-sm font-medium mb-1">
+              Due Date
+            </label>
+            <input
+              type="date"
+              value={dueDate}
+              onChange={(e) => setDueDate(e.target.value)}
+              className="w-full p-2 dark:text-gray-300 dark:bg-[#2D2D2D] border border-gray-300 rounded"
+              required
+            />
+          </div>
+
+          <div className="flex justify-end mt-4">
             <Button
               type="submit"
               disabled={loading}
@@ -354,7 +186,7 @@ const AddTask = ({
               {loading && (
                 <div className="w-5 h-5 border-2 border-white border-t-transparent border-solid rounded-full animate-spin"></div>
               )}
-              {isEditing ? "Update Task" : "Add Task"}
+              {editTask ? "Save Changes" : "Add Task"}
             </Button>
           </div>
         </form>
@@ -364,3 +196,4 @@ const AddTask = ({
 };
 
 export default AddTask;
+
